What I have figure out
- There is a need to implement the makeVarNames() in dataset_table.dart
in way to accomodate a lot of information about the dataset.

Meaning , the current implementation takes varNames as a parameters ,
but the implementation should be flexible enough to take in as many
parameters as the user wants.

For now though , there should be a way to show the types of the parameters